requirements.txt
Package                   Version   
------------------------- ----------
-yqt5-sip                 4.19.17   
-ywin32                   224       
2to3                      1.0       
altgraph                  0.16.1    
asn1crypto                0.24.0    
auto-py-to-exe            2.6.5     
bcrypt                    3.1.7     
bottle                    0.12.17   
bottle-websocket          0.2.9     
certifi                   2019.6.16 
cffi                      1.12.3    
chardet                   3.0.4     
Click                     7.0       
colorama                  0.4.1     
cryptography              2.7       
cx-Freeze                 6.0       
Eel                       0.10.4    
entrypoints               0.3       
escpos                    1.6       
future                    0.17.1    
gevent                    1.4.0     
gevent-websocket          0.10.1    
greenlet                  0.4.15    
hikvisionapi              0.2.1     
idna                      2.8       
jsonrpclib                0.1.7     
keyring                   19.0.2    
numpy                     1.18.1    
pandas                    1.0.0     
paramiko                  2.6.0     
pefile                    2019.4.18 
Pillow                    6.0.0     
pip                       20.0.2    
proteus                   5.2.2     
py2exe                    0.9.2.2   
pycparser                 2.19      
PyInstaller               3.5       
PyNaCl                    1.3.0     
pypiwin32                 223       
pypng                     0.0.20    
PyQRCode                  1.2.1     
PyQt5                     5.14.1    
PyQt5-sip                 12.7.1    
pyqt5-tools               5.11.3.1.4
PyQtWebEngine             5.14.0    
pyserial                  3.4       
python-dateutil           2.8.1     
python-dotenv             0.10.3    
pytz                      2019.3    
pyusb                     1.0.2     
pywin32                   224       
pywin32-ctypes            0.2.0     
qcrash                    0.2.0     
qrcode                    6.1       
random-password-generator 2.1.0     
requests                  2.22.0    
setuptools                39.0.1    
six                       1.12.0    
urllib3                   1.25.3    
whichcraft                0.6.0     
xmltodict                 0.12.0    
